Indenting and Debug mode
-indenting
-changed CanBeDismissedByTappingOutsideOfPopup = false on Debug mode

+++ Zaklad/ViewModel/ProductEditorViewModel_UserProduct.cs
@@ -14,10 +14,10 @@ namespace Zaklad.ViewModel
     {
         public ObservableCollection<Button> DecisionButtonsCollection { get; private set; }
         private IUserProduct _userProduct;
-        public IUserProduct UserProduct 
+        public IUserProduct UserProduct
         {
             get => _userProduct;
-            private set 
+            private set
             {
                 _userProduct = value;
                 OnPropertyChange(nameof(Gramature));
@@ -26,21 +26,68 @@ namespace Zaklad.ViewModel
                 OnPropertyChange(nameof(EditableFat));
                 OnPropertyChange(nameof(EditableCarbohydrates));
                 OnPropertyChange(nameof(ProductName));
-            } 
+            }
         }
 
-        public int Gramature { get => UserProduct.Gramature; set { UserProduct.Gramature = value; OnPropertyChange(nameof(Gramature)); } }
-        public decimal EditableKcal { get => UserProduct.Kcal; set { UserProduct.ProductTemplate.Kcal = GetProductTemplateValue(value); OnPropertyChange(nameof(EditableKcal)); } }
-        public decimal EditableProteins { get => UserProduct.Proteins; set { UserProduct.ProductTemplate.Proteins = GetProductTemplateValue(value); OnPropertyChange(nameof(EditableProteins)); } }
-        public decimal EditableFat { get => UserProduct.Fat; set { UserProduct.ProductTemplate.Fat = GetProductTemplateValue(value); OnPropertyChange(nameof(EditableFat)); } }
-        public decimal EditableCarbohydrates { get => UserProduct.Carbohydrates; set { UserProduct.ProductTemplate.Carbohydrates = GetProductTemplateValue(value); OnPropertyChange(nameof(EditableCarbohydrates)); } }
-        public string ProductName { get => UserProduct.Name; set { UserProduct.ProductTemplate.Name = value; OnPropertyChange(nameof(ProductName)); } }
+        public int Gramature 
+        { 
+            get => UserProduct.Gramature; 
+            set 
+            { 
+                UserProduct.Gramature = value; 
+                OnPropertyChange(nameof(Gramature)); 
+            } 
+        }
+        public decimal EditableKcal 
+        { 
+            get => UserProduct.Kcal; 
+            set 
+            { 
+                UserProduct.ProductTemplate.Kcal = GetProductTemplateValue(value); 
+                OnPropertyChange(nameof(EditableKcal)); 
+            } 
+        }
+        public decimal EditableProteins 
+        { 
+            get => UserProduct.Proteins; 
+            set 
+            { 
+                UserProduct.ProductTemplate.Proteins = GetProductTemplateValue(value); 
+                OnPropertyChange(nameof(EditableProteins)); 
+            } 
+        }
+        public decimal EditableFat 
+        { 
+            get => UserProduct.Fat; 
+            set 
+            { UserProduct.ProductTemplate.Fat = GetProductTemplateValue(value); 
+                OnPropertyChange(nameof(EditableFat)); 
+            } 
+        }
+        public decimal EditableCarbohydrates
+        {
+            get => UserProduct.Carbohydrates;
+            set
+            {
+                UserProduct.ProductTemplate.Carbohydrates = GetProductTemplateValue(value);
+                OnPropertyChange(nameof(EditableCarbohydrates));
+            }
+        }
+        public string ProductName
+        {
+            get => UserProduct.Name;
+            set
+            {
+                UserProduct.ProductTemplate.Name = value;
+                OnPropertyChange(nameof(ProductName));
+            }
+        }
 
         public ImageSource ProductImage => UserProduct.ProductImage;
 
         public event PropertyChangedEventHandler? PropertyChanged;
         private void OnPropertyChange(string name) => PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(name));
-        private decimal GetProductTemplateValue(decimal value) => value / (Gramature / 100);
+        private decimal GetProductTemplateValue(decimal value) => value / ((decimal)Gramature / 100);
         public ProductEditorViewModel_UserProduct(IUserProduct product)
         {
             UserProduct = product;
